Used options pattern to get configuration into file-based repositories
This still isn't great as e.g. the `FileBasedContactRepository` is now taking in data intended only for `FileBasedBranchRepository`, and it is not possible to have different delays for the different repositories.

We have, however, got rid of the horrible factory methods when setting up the DI for `IContactRepository` and `IBranchRepository` because they are no longer using primitives in their constructors. The downside of _that_ is it is a little harder to instantiate when unit testing.