Made options validation compatible with FluentValidation
This was made possible by sitting a little adapter class to turn FluentValidation's `AbstractValidator` into Microsoft's `IValidateOptions` interface. There is a base class provided so people can easily mix and match Data Annotations with Fluent Validation if they so choose to.

I don't much like the name `ValidateFluentValidation` but it was picked to follow the same format as ValidateDataAnnotations.