Fixed issue causing Fluent Validation of options to duplicate messages
We had registered multiple named options of the same type. This had resulted in the same type of validator being registered multiple times for that option. Microsoft's validation mechanism was then finding all registered validators for each instance of the option. As a consequence, if you had two named options you would see a failure message twice, if you had three named options you would see it three times, and so forth.

What this does show is if you register _one_ named option to be validated, then _all_ named options will be. I'm not yet sure if I consider that to be a problem.