Added extension method to insert a "record" into a cache
In this context a "record" is any type, rather than a type that is explicitly typed as record (although I just happen to be using a record in the tests). The data inserted into the cache is a JSON representation of the value.

This idea was inspired by Tim Corey in his [video on Redis](https://www.youtube.com/watch?v=UrQWii_kfIE)