Added ITypedCache and an implementation based on the Decorator pattern
It is not easy for people consuming our extension method to unit test. To do so they would have to do the same as we did i.e. test against the `GetAsync` and `SetAsync` methods on `IDistributedCache`. Consumers can now inject `ITypedCache` instead of `IDistributedCache` so `GetRecordAsync` and `SetRecordAsync` can be mocked.

I chose to leave the implementation as an extension method in `IDistributedCache` so if people wish to consume that interface they still have access to the method. The alternative would have been to have `ITypedCache` extend `IDistributedCache`.