Added Redis caching for IContactRepository using Decorator pattern
Implemented a decorator for the `IContactRepository` interface that attempts to retrieve the ContactEntity from a cache, before trying the repository it is decorating. We're using our own `ITypedCache`, rather than the `IDistributedCache` interface in order to make it easier to unit test our decorator.

The DI is configured so we're actually using a Redis cache behind the `ITypedCache` and we're still using the file-based repository behind the decorator for `IContactRepository`. This has led to the DI being a bit clumsy, and we would do with an extension method for easily decorating types.

This also includes configuring a secrets file so we're not exposing sensitive information like connection strings.