Implemented creating a contact using MediatR
There isn't really much going on here (yet) but this demonstrates it is fairly trivial to hook MediatR up to implement the Mediator pattern and support CQRS.

The requests and responses almost certainly need to move elsewhere, as do the handlers, probably into Application; I want to re-watch the SSW video on clean architecture before making too many decisions in that area.

It is nice to be able to use the model binding to have the command created for us, and not have to do anything.