Improved writing out of JSON
All operations on all instances of the `FileBasedContactRepository` were using the same options for serialization so, following Microsoft's guidelines, it made sense to share those, rather than recreate them on every call.

By setting `WriteIndented` to true we can keep the pretty format of the JSON file when writing to it. However, this seemed to break writing the `Title` and `Id` properties with lower-cased names, even when `PropertyNameCaseInsensitive` was true.

By introducing a very simple `JsonNamingPolicy` we can obviate the need for decorating properties of `ContactDto`, as well as fix the fact `Id` and `Title` were getting the wrong casing.