Changed Get for Contact to use Mediatr
You could _probably_ write the handler to accept a Guid, and return a ContactEntity as the response, to lessen the number of types we have floating around. However, we're planning on building this up to use pipelines to attach more things to both request and response...

This required modifying the `ValidateRequestPipelineBehaviour` to account for the fact we might _not_ have a validator for that specific request.