Added conversion operators for Guid <--> GetContactByIdQuery
I was hoping the model binding would be able to account for this, and I could now simply have the query type as the parameter to the action method. However that does not work, and we still have to create the query manually. As such, this doesn't really offer a great deal.

You could probably get the desired effect with a custom model binder, but that ends up being a lot of code just to move a Guid around!