Enhanced validation pipeline behaviour so we can return 400 BadRequest
In many ways this is nicer. We can now correctly return a 400 response when the command fails to validate. We can also return problem details as per [RFC 7807](https://tools.ietf.org/html/rfc7807). This also means we're not needlessly throwing exceptions which then clutter up the log output.

The downside is we've now got a very clumsy controller action that has to figure out what `IActionResult` to return, and we certainly wouldn't want to repeat that in every action method.

If the handler is not returning one of our special responses we'll still get the old exception-based behaviour.