Added pipeline behaviour for unhandled exceptions
This catches the exceptions, and adds them to one of our `Response` objects. This allows us to process all the exceptions in the same way, without every handler having to declare a `try ... catch` block and create the `Response` itself.

This shows a tidier way of applying generics so we can assign to the `Exception` property without having to mess with reflection. This can also be extended to work with the `ValidationResult`.