Tidied up public interface for Response and generic Response
We can do a little bit of reflection-based work to get at internal constructors, rather than going via public setters. That means we end up with a `Result` that cannot be manipulated once it comes out of the pipeline.

This has left us with a little bit of duplication between the two behaviours. Considering it is around reflection code this is not ideal!