Housekeeping of operations on Contacts
Experimenting with an idea inspired by Jimmy Bogard i.e. using an outer class to bring together the various moving parts for a single operation. This works reasonably well. Using the outer class means the consumer needs less qualification than if we used a namespace. I'm not convinced I like all the code in one file though as it is breaking the habit of a lifetime.