Extracted behaviour for converting Result into IActionResult
By turning this code into an extension method against Result we can make it reusable from any action. This seemed better than an extension method against `ControllerBase` as you would always have to qualify that with `this`. It also seemed better than introducing our own `ControllerBase` and forcing everyone to use that.

We explicitly did _not_ want this code on the `Response` class; we're intending for that to be usable from any type of application, and not tied into a HTTP context.

It would have been nice if we didn't have to pass the `ControllerBase` through, but trying to construct the `ObjectResult` for a validation problem was losing information. Even though I lifted an implementation from the implementation of `ControllerBase.ValidationProblem` I was losing information by virtue of not having access to the `ProblemDetailsFactory` 😢