Removed duplication of ResponseData from getting contacts
As this is a RESTful API it seems reasonable to return the actual entities as part of the MediatR integration. This means we don't need a specific `ResponseData` class for each query.

As we're returning entities from the GET action methods it seems reasonable to return `ActionResult<TValue>`, rather than `IActionResult` in order to help Swagger out.