Added IActionFilter to turn request validation exceptions into 400 responses
We were getting a nice 400 response _if_ the command handler returned one of our `Response` objects by virtue of the pipeline creating one with an exception. If the validation failed, but the handler returned something else, we would only get a 500 and an unhandled exception. This action filter detects that scenario and deals with it in a neater fashion.