Switched validation of CreateContactCommand to use MediatR
Previously we were using a now obsolete method of the `IBranchRepository` interface to determine if the branch existed. This was changed to use the new MediatR command. We _could_ also have still used the injected `IBranchRepository` and done our own querying against the `GetBranchesAsync` method. However, that behaviour is already implemented, so no need to re-invent the wheel, and this is also a nice demonstration that we're not restricted to just controllers when we lean on MediatR.