Fixed bugs relating to caching in CachingContactRepositoryDecorator
There were a couple of scenarios where not being able to interact with the cache could cause failures to be reported. This was especially nefarious in the case of an insert; the contact would be inserted, but we would still ultimately get a 500 response back. Fixing these issues means we're now in a position to start thinking about how to fit a circuit breaker into the decorator.