Changed IContactRepository to be able to return nulls when no contact matches passed ID
This is a bit of a cheat, but trying to hook Polly policies into the decorator was proving to be a major PITA due to the difference in nullability between getting something from the repository, and getting it from the cache. Ultimately we want to remove this method from the interface anyway, as we have done with `IBranchRepository`, so it isn't that big a deal.