Reimplemented CachingContactRepositoryDecorator using Polly policies
This is a bit of an abuse of the fallback policy IMO, which probably goes some way to explaining why it felt quite clumsy to get the implementation working. Fallback _feels_ like it should be for when we'd typically expect the operation to work. Here, we know there are a great many occasions where our first attempt will not work.

Another downside here is we're now coupled to a Redis implementation by virtue of the policies in the decorator having to handle a specific exception type.

Having said all that, this was done as a learning exercise and it has served its purpose in that regard. It also opens up the option of us now adding in a circuit breaker, timeout and various other things, also as a learning exercise.