Added options to allow configuration of circuit breaker in CachingContactRepositoryDecorator
Because this type is scoped as transient each instance is getting its own policy wrapper. This means that when the full application runs a circuit breaker never trips because we don't know about failed executions from other requests!

We've got a few options (no pun intended) for fixing this. We could either scope the decorator as singleton which wouldn't be terrible. We could configure the policy as static to the class, so even the type is scoped as transient we would share the policy. We could also share the policy by using a policy registry and getting the policy from there. This would, however, necessitate a split in how the policies are configured; realistically can't put the fallback policies in the policy registry because they need access to things that should not be public!