Experimented with the Polly caching policy to explore behaviour around exceptions
This has taught me a bit more about how the various providers work if I have to set them up outside of the typical DI patterns. It has also allowed me to figure out some non-trivial uses of Moq to configure more sophisticated behaviours. Although this code doesn't add anything to the application it is useful reference material.

This has confirmed my suspicion that the cache and circuit breaker policies won't play nicely together if e.g. a Redis server is not contactable. In that scenario the call to the cache is very slow as we wait for an exception to be thrown. However, due to the cache swallowing those exceptions the circuit breaker never trips. As such the cache policy will actually slow you down if your Redis server is FUBAR. This will need all of my super-powers to resolve, but I have a cunning plan...