Added test to prove we can cache and retrieve a value using our new policy
This is the standard happy case scenario. The configuration of the tests feels quite clunky, but I don't (yet) see a way to simplify that. We've got a lot of moving parts such as caches, key strategies, and TTL strategies. We need some non-trivial configuration of the mocks the SUT uses. We need to execute things through the SUT, and we're using mocks to check their execution, but we also need non-trivial configuration of those mocks too!